First commit of signal from aio branch with aio removed
