+++ posixlib/src/main/resources/scala-native/signal.h
@@ -1,0 +1,9 @@
+#ifndef SIGNAL_H
+#define SIGNAL_H
+
+// aio uses sigevent and sigval
+//struct scalanative_sigval {
+//
+//};
+
+#endif
